requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
influxdb==5.3.1
msgpack==1.0.2
psutil==5.8.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
urllib3==1.26.7
